Move query filter-related code to models/query because it existed first
